+++ src/app/components/constitution-page/votes/votes-grade/votes-grade.component.ts
@@ -5,11 +5,10 @@ import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { CARDS_SORT_KEY, CARDS_VIEW_KEY, GRADE_SHOW_STATS_KEY, GRADE_ALREADY_VOTES_KEY } from 'src/app/types/local-storage';
 import { mean, variance } from 'src/app/types/math';
-import { compareObjectsFactory } from 'src/app/types/song';
+import { compareObjectsFactory, toMap, toMapNumber } from 'src/app/types/utils';
 import { getEmbedURL, getIDFromURL } from 'src/app/types/url';
 import { VoteNavigatorComponent } from './vote-navigator/vote-navigator.component';
 import { ActivatedRoute } from '@angular/router';
-import { toMap, toMapNumber } from 'src/app/types/utils';
 import { YatgaUserFavorites } from 'src/app/types/extends/favorite';
 
 enum GradeOrder {
